feat: segment footer
