+++ packages/splits-kit/src/components/DisplaySplit/index.tsx
@@ -8,7 +8,6 @@ import SplitRecipients from '../DisplaySplit/SplitRecipients'
 import SkeletonLoader from '../DisplaySplit/SkeletonLoader'
 import SplitBalances from '../DisplaySplit/SplitBalances'
 import SplitHeader from '../DisplaySplit/SplitHeader'
-import ComponentLayout from '../util/ComponentLayout'
 import ChainLogo from '../util/ChainLogo'
 import { IAddress } from '../../types'
 import Segment from '../util/Segment'
@@ -54,58 +53,55 @@ const DisplaySplit = ({
     metadataError && metadataError.name === 'AccountNotFoundError'
 
   return (
-    <ComponentLayout width={width}>
-      <Segment
-        title={<SplitHeader address={address} />}
-        titleButton={
-          <div className="flex items-center space-x-2">
-            <Link
-              href={`https://app.splits.org/accounts/${address}/?chainId=${chainId}`}
-              className="cursor-pointer text-gray-500 transition hover:text-black focus:outline-none dark:hover:text-white"
-            >
-              <ArrowTopRightOnSquareIcon className="h-4 w-4" />
-            </Link>
-          </div>
-        }
-        corner={
-          displayChain
-            ? CHAIN_INFO[chainId] && (
-                <ChainLogo chainInfo={CHAIN_INFO[chainId]} />
-              )
-            : undefined
-        }
-        body={
-          isNotFound ? (
-            <div className="text-center mt-12 space-y-2">
-              <div className="text-lg">Split not found</div>
-              <div className="text-xs">
-                This account is not a Splits contract on the{' '}
-                {CHAIN_INFO[chainId].label} network.
-              </div>
-            </div>
-          ) : (
+    <Segment
+      width={width}
+      title={<SplitHeader address={address} />}
+      titleButton={
+        <div className="flex items-center space-x-2">
+          <Link
+            href={`https://app.splits.org/accounts/${address}/?chainId=${chainId}`}
+            className="cursor-pointer text-gray-500 transition hover:text-black focus:outline-none dark:hover:text-white"
+          >
+            <ArrowTopRightOnSquareIcon className="h-4 w-4" />
+          </Link>
+        </div>
+      }
+      corner={
+        displayChain
+          ? CHAIN_INFO[chainId] && <ChainLogo chainInfo={CHAIN_INFO[chainId]} />
+          : undefined
+      }
+      body={
+        isNotFound ? (
+          <div className="text-center mt-12 space-y-2">
+            <div className="text-lg">Split not found</div>
             <div className="text-xs">
-              {isLoadingMetadata || isLoadingEarnings ? (
-                <SkeletonLoader />
-              ) : (
-                <div className="space-y-4">
-                  <SplitRecipients split={split} />
-                  {displayBalances && !isLoadingEarnings && (
-                    <SplitBalances
-                      chainId={chainId}
-                      address={address}
-                      formattedSplitEarnings={formattedSplitEarnings}
-                      onSuccess={onSuccess}
-                      onError={onError}
-                    />
-                  )}
-                </div>
-              )}
+              This account is not a Splits contract on the{' '}
+              {CHAIN_INFO[chainId].label} network.
             </div>
-          )
-        }
-      />
-    </ComponentLayout>
+          </div>
+        ) : (
+          <div className="text-xs">
+            {isLoadingMetadata || isLoadingEarnings ? (
+              <SkeletonLoader />
+            ) : (
+              <div className="space-y-4">
+                <SplitRecipients split={split} />
+                {displayBalances && !isLoadingEarnings && (
+                  <SplitBalances
+                    chainId={chainId}
+                    address={address}
+                    formattedSplitEarnings={formattedSplitEarnings}
+                    onSuccess={onSuccess}
+                    onError={onError}
+                  />
+                )}
+              </div>
+            )}
+          </div>
+        )
+      }
+    />
   )
 }
